PieChart hotfix to allign segments in circle

+++ main/PieChart.pde
@@ -30,15 +30,15 @@ class PieChart{
 
 class PieChart {
   float diameter;
-  int[] data;
+  float[] data;
   int dataTotal;
   
-  PieChart(float diameter, int[] data) {
+  PieChart(float diameter, float[] data) {
     this.diameter = diameter; 
     this.data = data;
   }
   
-  void pieChart(float diameter, int[] data) {
+  void pieChart(float diameter, float[] data) {
     float lastAngle = 0;
     
     color[] colors = {color(255, 0, 0), color(0, 255, 0), color(0, 0, 255), color(255, 255, 0), color(0, 255, 255), color(255, 0, 255), color(128, 0, 0), color(0, 128, 0), color(0, 0, 128), color(128, 128, 0),
